Fully implement ICMP4 echo request / reply.
The identifier and sequence number are part of the 32-bit ICMP header data field(which was missing from the struct definition), not of the payload.  Partly as a consequence of this misunderstanding, the payload position and length were both off by four, and the identifier and sequence number were being overwritten by the payload.

In addition, both the ICMP and IP checksums were computed with garbage in the checksum field instead of zero, due to using malloc() instead of calloc().