Fix ARP counting / timing logic.
Fix the timing logic by not updating the last-seen stamp until after we've inspected it.  This breaks the expiry code, but it was never used, so #if it out for now.

Change the trigger condition to claim an address after, rather than upon, the third unanswered ARP request (i.e. upon the fourth).

Add an upper time limit: if more than 30 seconds have passed since the last unanswered ARP request for an address, start over as if we'd never seen it before.