Implement tarpitting.
Upon receipt of a naked SYN, we respond with a SYN+ACK with a window size of 0, while packets containing data are answered with naked ACKs with the correct sequence numbers but still 0 as window size.  This causes the sender to retry periodically until something times out.  Since we don't store state for the connection, we can go on indefinitely.

We also respond to SYN+ACK (with RST) but not to FIN, because answering FIN correctly requires keeping connection state.