Fix Travis build with -Wshadow.
Linux's bswap() implementation uses non-portable pseudo-closures with local variables, meaning that the argument to bswap() can't be an expression that calls bswap(), as the local variables in the inner closure will shadow those in the outer closure.  Work around this by using a temporary variable, which the compiler will hopefully optimize out.