Fix removal of unaligned ranges.
When splitting a leaf before removing a sub-range of that leaf, we need to create child nodes for the subtrees that will be left after the removal.  Because we only looked at fsub and lsub, we would only create child nodes for subtrees that would survive in their entirety, not for subtrees which would later be partly removed.  Note that we could just create a full set and then remove what we didn't need, but I don't like unnecessary allocations.