Add destination filtering.
The command-line syntax, implementation, and documentation are copied (with modifications) from flytrap(8).  Note that this means inverting the meaning of -i / -x, which applied to the source address in one but to the destination address in the other.  Breaking compatibility for a tool that no-one (probably) uses is seen as a lesser evil than having two tools which accept the same syntax with diametrically opposite semantics.

While here, harmonize the usage string and documentation for both flytrap(8) and ft2dshield(1).